CR1000937: Avoid updating specific configuration records on package update

+++ api/include/SpiceUI/SpiceUIConfLoader.php
@@ -84,6 +84,31 @@ class SpiceUIConfLoader
         'sysuirolemodules',
         'sysuiroles',
         'sysuiroutes',
+        'sysexchangemappingsegments',
+        'sysexchangemappingsegmentitems',
+        'sysexchangemappingmodules',
+        'sysmsgraphmappingsegments',
+        'sysmsgraphmappingsegmentitems',
+        'sysmsgraphmappingmodules',
+    ];
+
+    /**
+     * specific records from tables without package column shall be inserted, NOT updated, NOT deleted
+     * @var string[]
+     */
+    private $insertOnlyTables = [
+        'nodata',
+        'sysfts',
+        'sysnumberranges',
+        'sysnumberrangeallocation',
+        'syssalesdocnumberranges',
+        'syssalesdoctypes',
+        'syssalesdoctypesflow',
+        'syssalesdoctypesitemtypes',
+        'syscategorytrees',
+        'syscategorytreelinks',
+        'schedulerjobtasks',
+        'schedulerjobs'
     ];
 
     /**
@@ -190,11 +215,9 @@ class SpiceUIConfLoader
                 $tables[$conftable][] = $recordId;
             }
         }
-        // skip Tables
-        $skipTables = ['nodata', 'sysfts', 'sysnumberranges', 'sysnumberrangeallocation'];
 
         foreach ($tables as $conftable => $recordIds){
-            if(in_array($conftable, $skipTables)){
+            if(in_array($conftable, $this->insertOnlyTables)){
                 continue;
             }
             try {
@@ -355,11 +378,28 @@ class SpiceUIConfLoader
                     foreach ($decodeData as $key => $value) {
                         $decodeData[$key] = (is_null($value) || $value === "" ? NULL :  $value);
                     }
+
+                    // set the flag to check on insert
+                    $skipInsert = false;
+
                     //delete before insert
-                    $delWhere = ['id' => $decodeData['id']];
-                    if(!$db->deleteQuery($tb, $delWhere)){
-                        LoggerManager::getLogger()->fatal("error deleting entry {$decodeData['id']} ".$db->lastError());
+                    if(!in_array($tb, $this->insertOnlyTables)){
+                        $delWhere = ['id' => $decodeData['id']];
+                        if(!$db->deleteQuery($tb, $delWhere)){
+                            LoggerManager::getLogger()->fatal("error deleting $tb entry {$decodeData['id']} ".$db->lastError());
+                        }
+                    } else{
+                        // check if record is present
+                        if($dbResRow = $db->getOne("select * from $tb where id='{$decodeData['id']}'")){
+                            $skipInsert = true;
+                        }
                     }
+
+                    // skip insert if this is a record we should keep
+                    if($skipInsert) {
+                        continue;
+                    }
+
                     //run insert
 //                if($tb == 'email_templates'){
 //                    file_put_contents('spicecrm.log', 'dict email_templates '.print_r(SpiceDictionaryHandler::getInstance()->dictionary['EmailTemplate'], true)."\n", FILE_APPEND);
